pyproject.toml
[project]
name = "news-didattica"
version = "2.0.0"
description = "A Telegram bot to scrape and send news from the Didattica website."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4==4.12.2",
    "markdownify>=1.1.0",
    "numpy>=2.0.2",
    "openpyxl>=3.1.5",
    "pandas==2.2.3",
    "python-dotenv>=1.1.0",
    "python-telegram-bot==20.5",
    "requests==2.31.0",
]
